JbdcAccessMethod: add test method with big table
This test simply uses JdbcQueryScan to 'select * from testtablebig' and
count up the number of results. The connection IP is the elastic IP of
the AWS micro instance hosting the MySQL myriad_test database, and the
other credentials are hardcoded in the Java file.

The database itself is configured using the sql/aws_mysql_test scripts,
which are also updated to setup the testtablebig. reset_db.sh must be
run at least once before this test will pass. (As of this commit, it has
been run.)

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>