JdbcAccessMethod: cache connection objects
Creating a JDBC connection is a pretty heavyweight operation. If we're
scanning multiple tables, the operators currently create a new
connection for each QueryScan, which thus results in lots of network
overhead.

Instead, cache each connection in a static object and reuse if
appropriate.

In this implementation, a cached JDBC connection will be reused if the
String parameters driverClassName and connectionString are equal.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>