Created the test to JdbcAccessMethod that times the write of tuple batches.
Moved the JdbcAccessMethodTiming test to the speedtest package.

Updated speed test for Jdbc remote connection.

Created the test to JdbcAccessMethod that times the write of tuple batches.

removed the duplicate timing test file -- it was moved to the speedtest

updated new changes to speed test for remote sql (on aws) insertions.

Squashing multiple commits on speed test for insertions into a remote mysql database (using test db on aws)