Producer: use IPCConnectionPool instead of Worker
The Producer operators could not be used at the Server because they
required a pointer to a Worker. But actually, they only used a Worker to
get access to its connection pool. So, store the ConnectionPool instead
of the Worker and then it can also be used at the Server.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>