ShuffleConsumer: don't require a child ShuffleProducer
Sometimes, we may want to have a ShuffleConsumer in a query plan
separate from a ShuffleProducer, i.e., if a node is acting as a
'reducer' (receiving data) but not also as a 'mapper' (sending data).

Add another constructor that takes a Schema instead of a child operator,
and then make sure the code works in the cases where it is constructed
with Schema but then the child operator is set.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>