myriad: make SQLite use the same database for all tables
If we're going to actually make use of the database, we need to store
multiple tables in each file. Convert the catalog and worker over to
this new storage technique, and fix a few bugs that cropped up as a
result.

Note that to make sure that ShuffleSQLiteTest worked I had to introduce
a busy-timeout in SQLiteAccessMethod. This is necessary because SQLite
does the "whole-database" lock. This may quickly lead to problems for
big datasets and might necessitate using a different database system
with finer-grained locking.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>