myria: changes to enable MonetDB
1) MonetDB doesn't let us use [user#program#relation] as the relation
key; the only sensible escape I can come up with for a tri-partite key
is "user program relation". This requires us to refactor the way we use
RelationKey all over the place---we can't just always use its toString
operator. Now we also need to change escape ('[' vs '"') and separator (' '
vs '#') characters.

1a) For error-checking, make RelationKey.toString() throw an
IllegalArgumentException. This way we can catch places we missed using
it quickly.

2) Add JdbcInfo as a container for all the JDBC connection parameters.
This is a useful way to abstract different JDBC drivers out of the
system and cleans up some of the code.

3) Add tests for MonetDB access and speed.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>