myriadeploy: fix error-handling and syntax in scripts
Instead of doing `cmd1 ; cmd2` we really want to do `cmd1 && cmd2` so
that `cmd2` is not executed when `cmd1` fails. This also means that
better error codes will get passed to the script.

(What was happening: the `cd` failed and then `startMaster.py`, etc.,
were creating files in other directories.)

Also, Python doesn't use `;` at the end of lines :)

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>