myriadeploy: don't ssh or remote rsync when hostname='localhost'
Turns out CSE support has a cap on the number of times you can log into
a machine within 1 minute; going over this limit gets you banned for 30
minutes. It's easy to hit this ban when setting up a Myria cluster on
localhost--you just need 3 workers or more.

As a workaround, don't use ssh/remote rsync if hostname is localhost.
This also means we can't change usernames, so put a note in the
deployment files about that problem as well.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>