FileScan: properly handle unix and windows line endings
There was an issue with Windows line endings (#153). Fix by updating the
delimiter regex in FileScan.java. Also added a new testdata file and
unit test.

Since git automagically corrects line endings when files are checked
out, set the git attributes for the original test to be UNIX line
endings (LF) and the git attributes for the new file to be DOS line
endings (CRLF).

Close #153.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>