Encoding: refactor QueryEncoding to the annotated plan fragments
Queries now take the form of annotated plan fragments:

	query = {
	  'raw_datalog' : 'blah',
	  'logical_ra' : 'blah',
	  'fragments' : [ fragment1, fragment2, ... ]
	}

A fragment is simple a 2-element dictionary containing a list of
`operators` in query fragment and an **optional** list of `workers` to
execute that fragment.

	fragment = {
	  'operators' : [ operator1, operator2, ... ],
	  'workers' : [ 1, 2, 3, ... ]    # OPTIONAL
	}

As before, an operator is a dictionary that has an `op_type`, an
`op_name`, and optional arguments:

	operator = {
	  'op_type' : 'SQLiteScan',
	  'op_name' : 'Scan0',
	  ...
	}

Here's the new hotness!

1. In Producers and Consumers, the old `operator_id` integers that map
to `ExchangePairID`s inside of Myria are now replaced with String
descriptions. For example, for the `ShuffleProducer` and
`ShuffleConsumer` pair that send/receive a tuple hashed on its first
attribute, we could label its `operator_id` "hash(x)". Hopefully this
makes the plans easier to read. This also means that we don't have to
carefully allocate numbers; we can just assign them distinct strings.

2. Producers and Consumers also no longer let you configure the actual
worker IDs of the workers that will run them. We instead keep track of
which fragments contain the Producers & Consumers for each `operator_id`
and then configure the Producer/Consumer children using the
Consumer/Producer maps. This should be less error-prone!

3. Finally, we provide the option to the user (query submitter) of
letting the system decide which workers should run each fragment. The
rules are currently very simple:

    a. If the fragment contains a scan, go figure out which workers
    store that relation. Use those workers for this fragment.

    b. Otherwise, use all living workers for this fragment.

    If you provide a `workers` field in your fragment, these rules will
    be ignored.

    There is not currently functionality to automatically detect
    singleton operators like `EOSController` and only use one of them.
    Though, perhaps we could add these rules in the future?

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>