Fix take() call blocks offer()
The take() in FlowControlInputBuffer implementation was incorrect. One
thread calls take() will block all other threads from calling offer().
Therefore, the whole system will be blocked forever. The reason is
that the take() method waits on the storage's inner lock, which is out
of our control. Now the new version waits on the eventSerializeLock
which is under our control.