Specify worker nodes to ingest data
Data worker nodes are now decoupled from worker compute nodes. It is possible to, in the ingest json file, specify a set of worker nodes to ingest data into. This set should obviously be a subset of the worker nodes. If the worker nodes list is omitted, than all alive worker nodes will be used, so that retrocompatibility to all older files is kept.