Added two json sample files for data ingestion and a join query.
The data ingestion json file, only worker nodes 1 and 2 will receive data. In the query plan in the join json file, the worker nodes 1 and 2 perform scan on the files they have and shuffle them to the 4 nodes on the deployment; all 4 nodes receive the data, perform a local join, and send the results to the master.