Specify path per node in the deployment config file.
The path for deploying files is not unique for the whole system anymore. One can now specify a path for each node, the master and all workers. Now, even workers in the same node can use different directories, which means that we can use parallellism inside a node, with different disks (and disk controllers).

The file myriadeploy/deployment.cfg.local maintains all nodes with the same directory.

On the other hand, the file myriadeploy/deployment.cfg.sample uses three machines, vega as master; aldebaran and altair as workers. The four disks we have on those worker machines are being used in this sample deployment configuration.