api.encoding: simplify the MyriaApiEncoding required field interface
In the old version of the code, every class extending `MyriaApiEncoding`
needed to implement its own `validate` logic that basically just checked
that the required fields were non-null, was required to call
`super.validate()`, and throw a hand-generated exception following the
same arguments.

Dramatically simplify this by moving all logic to ensure required fields
are present to `MyriaApiEncoding.validate()`; extending classes now need
only define a function that returns the list of names of fields they
require to be present. For operators that need further validation, they
can extend the adaptor-style method `validateExtra()` from
`MyriaApiEncoding`. This is done, for instance, by
`ShuffleProducerEncoding`, which checks that its partition function is
valid.

Also fix the bugs in the operators that weren't validating their fields,
such as `QueryEncoding` or `ShuffleProducerEncoding`.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>