Use TupleSource as test data source
SQLiteConnection confines to the creation thread. The current
nonblocking executor uses a thread pool to execute the tasks. It may
result in Exceptions if the SQLiteScan operator is executed in
different threads in the thread pool. This is a common issue. A more
sophsicated nonblocking executor should be designed to handle the
thread confinement stuff.