Define clearly a streamID
A stream is identified by a pair (remoteMachineID, streamID).
remoteMachineID is an integer, i.e. the IPC entity ID. streamID is a
long. The StreamOutputChannel with an ID sends data to the
StreamInputChannel with the same ID. The StreamID is an extension to
the old ExchangeChannelID, which identifies a data stream by only a
long operatorID.