Add StreamInputBuffer and ShortMessageProcessor
Add StreamInputBuffer and ShortMessageProcessor.

StreamInputBuffer serves as the root class of all stream data buffer.
A StreamInputBuffer has a set of StreamInputChannels as it's inputs.
An input buffer buffers data from its inputs and also maintains the
BOS and EOS states of its input channels. The input buffer starts to
receive data after it's get started (call start) and stops when all
the input channels are EOS.

The ShortMessageProcessor is the root class of all short message
processors. All the data which are sent by
IPCConnectionPool.sendShortMessage in any worker to a given worker is
processed by an implementation of ShortMessageProcessor at the
receiving worker.