Implement a new flow contorl input buffer
1) Implement a new FlowControlBagInputBuffer using the new
StreamInputBuffer interface. The old InputBuffer and
FlowControlInputBuffer are abandoned.

2) add EOI into TupleBatch. The EOI part is used only in Consumer for
identifying if a TB payload from IPC layer is an EOI. The operator EOI
implementations are still using the Operator.eoi boolean.

3) The new FlowControlBagInputBuffer adds a new type of recover
envent. The recover event will be triggered if a app-specific recover
trigger is meet. The recover trigger is a number less than the size of
the input buffer. If the previous event is buffer full and the number
of messages in the input buffer is less than the trigger, the recover
event is triggered.