Add PayloadSerializer
A payload serializer is the place where the serialized binary data and
the java objects gets converted to each other. In Myria, we use the
TransportMessageSerializer. For data messages, when to serialize, it
serialize the TupleBatches to TransportMessage and then to binary
array (ChannelBuffer). And also when to deserialize, it converts the
binary array data back to TransportMessage and then back to
TupleBatch.