Changed some constants in order to be generic in terms of database system
Changed:

* WORKER_STORAGE_SYSTEM_TYPE   --> WORKER_STORAGE_DATABASE_SYSTEM
This constant contains the database system, e.g. "sqlite", "monetdb", "mysql", or "vertica"

* WORKER_DATA_SQLITE_DB        --> WORKER_STORAGE_DATABASE_NAME
This constant contains the database name, which for SQLite is the file, but for others can be a directory

* EXEC_ENV_VAR_DATABASE_SYSTEM --> (created)
* EXEC_ENV_VAR_SQLITE_DB       --> EXEC_ENV_VAR_DATABASE_NAME
These two will be used to send the values of database system and name to the operators through the execEnvVars variable.