LocalJoin: change constructor from Schema to List<String>
As the code has changed the code over time, we (@jingjingwang mainly,
also me) figured out that the "best" way to use LocalJoin is to not let
the user pass in the Schema: it's too easy to get it wrong, and we can
figure out column types and names from the childs' Schemas.

The downside is we can't change column names at the join, which we
probably want to be able to do. So, add a column names parameter and use
it.

Also, the old code used a hacky check in Schema.merge to automatically
resolve column name duplicates. (It only worked if left[i] == right[i]).
The new code does not call this function anymore, which means that we
need to pass in explicit names, and I fixed all unit tests that had this
issue.

We could also fix this by including a more robust version of
Schema.mergeAndRename() e.g. that renames all columns as they come up.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>