Add 3 StreamingStateUpdater: DupElim, KeepMinValue, SimpleAppender.
DupElim used to be an operator but not anymore. It's a
StreamingStateUpdater that should be attached to an operator which
implements StreamingAggregate, for example, an IDBInput or a
StreamingAggregateAdaptor.

KeepMinValue always keeps the tuple with smaller value under the same
key.

SimpleAppender simply just keep appending tuples with no replacement.