systemtests: fail individual tests after 120s
Usually the system tests complete in ~10 minutes on my computer, and
sometimes I have to kill them after 30 minutes.

These are the tests that, on my computer, often don't pass after 60s:

    FTModeTest. abandonTest
    FTModeTest. rejoinTest
    FlowControlTest. flowControlTest
    QueryFailureTest. workerInitFailureTest

There appear to be some bugs related to flow control and/or fault
tolerance and/or query failure.

Add a 120s time limit (per test case) to all system tests, so that the
test suite will eventually complete. We might even try and bring this
limit down as we try to get the system into better working shape.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>