Move the "setting WAL mode" part from SQLiteAccessMethod to Insert & Scan operators.
Commit 9f81c3a made twitterSubsetProjectingJoinWithInsertTest pass but
tests in TwitterJoinSpeedTest fail, the reason, I believe, is still:
when two threads are setting WAL mode on the same db file, they both
require exclusive lock and then lead to a confliction. The solution is
to ensure that for each db file, WAL is set in the beginning by only one
thread, and putting the procedure in Operator.init() can solve the
problem since all the operators' init() methods are called one by one, in one thread.