Operator: cache schema in the Operator, regenerate on demand
We've had a series of bugs in most operators where the schema is not
prepared when we need it. See 7ceb9c7, 79d71a2, 050bf75, b39cf9a.

Refactor the way that schemas are generated in a more general way to
prevent this problem from arising in the future. Then fix every Operator
in the system to use the new method.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>