AccessMethod/DbInsert: fix logic around overwrite/create/drop tables
Recent changes to AccessMethod made it completely ignore the flags that
controlled whether existing datasets were appended-to or deleted.
Refactor the entire logic around DbInsert and the relevant functions in
the access methods to handle this.

We now require the following DDL functionality from any DBMS:

* CREATE TABLE IF NOT EXISTS
* DROP TABLE IF EXISTS
* RENAME TABLE

MonetDB does not support standard SQL DDL commands, so it complicates
a lot of the logic around creating (and overwriting) tables. Since we've
decided we're not using MonetDB as a standard DBMS, let's just drop it.

So I disabled the MonetDB test. If someone wants to use MonetDB, you can
update the JdbcAccessMethod to implement the multi-command versions that
might be able to support these in MonetDB.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>