UnionAll: fix semantics
There was an occasional bug that happend if the last child in the list
was EOS. We got the child via next(), we then regenerated the list, and
then we called itr.remove(). This was either a different child (the
first one in the list), or an IllegalStateException (if there are now no
children with data).

Fix this bug by dramatically simplifying the logic. Rather than keeping
a list of integers and generating iterators to simulate a circularly
linked list, just actually use a circularly linked list by appending
children on the end, when appropriate.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>