Make sure that we get an EOI on both sides if we get an EOI on one side and know that we don't need to proceed on the other child.
This is different in merge join because we are able to know that we don't output any tuples any more even though we haven't drained both children. For EOS, we can simply return EOS but for EOI we have to drain because we need to start at the first tuple batch in the next iteration.