Column/TupleBatch: make Column physical layout independent of TupleBatch
In other words, from the outside, every Column should look like it is
"dense". This dramatically simplifies a lot of APIs (e.g., none of this
indirecting through validIndices except when necessary), especially
inside of TupleBatch and TupleBatchBuffer.

One reason to do this is to make it much easier (and lightweight) to
manipulate TupleBatches. For instance, suppose I want to append a column
to an existing TupleBatch. If there has been a filter applied to that
TupleBatch, I have to compact it into a new TupleBatchBuffer! This
should be a very lightweight operation yet is complicated and
computationally expensive.

We can also do things like have a "ConstantValueColumn" that only stores
a single value and for which filter is almost a no-op. We can
encapsulate Columns in other Columns. Etc.

I replaced the TupleBatch-wide validIndices framework with a
FilteredColumn class that wraps an inner column in a filter. I then
propagated all the resulting changes to TupleBatch, TupleBatchBuffer,
etc...

Along the way I fixed a bug, or maybe just an idiosyncracy, in the way
that EOI tuplebatches are handled. All TupleBatch objects are required
to have a number of columns equal to the size of their schema, yet the
EOI tuple batches did not. This led to some strange and non-obvious
semantics in the handling of these batches in TupleBatchBuffer!

I think that code is much clearer now, without these hidden invariant
violations.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>