Add a second kind of listener to OperationFuture
Add pre-notify listeners to OperationFuture. These listeners get
executed after the operation is done and before waiting threads
getting woke up. This is to solve existing race conditions. For
example, in DatasetMetadataUpdater, the metada gets updated after a
qurey is executed by getting executed in a QueryFuture listener. But
if any thread is syncing on the same QueryFuture, the thread may see
  the query is finished but the metadata is still not updated. Meet
  this race condition in a run of systemtest.JsonQuerySubmitTest