TimerTaskThreadFactory is wrong
The TimerTaskThreadFactory was trying to prevent timer tasks from been
unscheduled when error thrown, which is the behaviour in the
ScheduledExecutorService. But the implementation is wrong. The correct
implementation is to capture errors for each submitted task rather
than the whole backend executing process (wich is implemented as the
TimerTaskThreadFactory).

ErrorLoggingTimerTask is the correct implementation.