Disable ShutdownThreadCleaner
Move (backup) ShutdownThreadCleaner to util and disable it in Worker.
This class forcely kill threads when shutdown if the threads are not
willing to shutdown themselves. But it actually cover several
threading bugs. For example, an unexpected deadlock occured between
two tasks when they try to kill each other at failures. The query
execution still succeeds because a query gets marked failure when any
single task reports failure. The two deadlock threads will get killed
by the ShutdownThreadCleaner when Worker shuts down.

The deadlock problem will be fixed in a later commit.