Worker,Server: update workaround #371 to fix deadlock
To work around #371, I added memory limits to channels. However, I made
the per-channel and the global memory limits the same. This is clearly
nuts: multiple channels can be active at the same time!

In some plans with multiple fragments that produce data, I believe that
one it is possible that one channel consumes all the memory and
deadlocks the query execution because that channel's buffer won't be
emptied until another channel on the same work is pushed.

Fix this by getting rid of the global limit.