StatefulApplyTest: fix & test for bug #547
Fix GenericEvaluator for StatefulApply. Basically, the old code doesn't
let you compile expressions that copy from the child input or produce
constants (!needsCompiling()) as a way of preventing the caller from
using the tuple-wise eval method when they should be processing a
TupleBatch in O(1) steps.

However, for StatefulApply's updateExpressions we need to run them
row-at-a-time and hence we must compile them. Enable this by updating
the guard condition in GenericEvaluator.compile() from
`needsCompiling()` to `needsCompiling() || (getStateSchema() != null)`
so that it can be compiled for StatefulApply.

Also add a test of a simple StatefulApply that simply copies its input.