parallel: add RelationWriteMetadata
When a relation is written to by a query, this class is used to
determine which workers will contain the relation, what the Schema is,
whether an existing relation by that name will be overwritten, etc.
The information is kept consistent across workers for every subquery
e.g., if two different workers write a relation of the same name, it must
have the same Schema in both cases and both workers must have the same
value for the overwrite bit.