DatasetMetadataUpdater: refactor
Move the functionality for inferring the relations created to SubQuery.

Keep track of whether the relations are being overwritten or not in the metadata.

Only make sure the schema matches the catalog if the overwrite bit is false.

If the overwrite is true, delete all traces of the existing relation on query success.