add DoWhile support to Myria
The DoWhile operator is like a Sequence, except that it takes an
additional parameter: the name of a singleton boolean relation (1
column, 1 row, must be a boolean) that determines whether the entire
body should be run again.

I believe that DoWhile operators can be nested, but as raco will not
emit these plans this is not yet tested.

The DoWhile is not especially well sanity-checked. E.g., we don't check
that the "loop condition" relation is actually written in the body. User
beware --- as we write more programs that use it we will undoubtedly
find better places to put checks.