Query: add global variables and the ability to set them
Each query now has a global variable store. This might be used in a
number of ways, e.g., to fill in parameters of later queries. At the
moment, I want it to hold the loop condition for DoWhile.

Additionally, I added accessors to the global variables through Query
and Server.

I also added an Operator called SetGlobal that sets the global variable,
and encodings of it. The Operator is tested.