myria: make logger.trace more efficient and remove unnecessary checks
1. Building strings is wrong. Use the LOGGER.x("{}, {}", a, b) notation instead
of LOGGER.x(a+", "+b). More more efficient!

2. Since the latter is much more efficient, in most cases we do not need
LOGGER.isXEnabled before it. In particular, when a and b are passed by value as
objects, there is absolutely no savings from the check. Only where a or b are
constructed objects or the result of heavyweight function calls is the check
worthwhile.