Fix another cause: remove pipeline executors.
The pipeline executors have their own memory management mechnism by
specifying the estimated size of message caches. However, they can not
work together with FlowControlInputBuffer or any other backpressure
mechnisms.

The reason is that Netty provides a boolen switch for each connection
to turn on/off of data transmission. The executors and the
FlowControlInputBuffers are both using this switch to implement
backpressure. If we have both the executors and the
FlowControlInputBuffers present, we have three layers of threads:

Netty worker threads --> pipeline executor --> myria query executor

--> here means push message into

When the myria query executor finds the input buffer is too full, and
turn the connection switch to stop read data, the pipeline executor
does not know that, it continues push data into input buffer and
restart the network data transmission.

Currently what the pipeline executors do are justing
serializing/deserializing messages, removing them should not hurt the
performance.