Fix startAProcess
simply split a string is not correct. In "ssh address shell_commands",
the third argument should be a single string. The old code will split
the shell_commands to an array of strings and serves into
ProcessBuilder, which is not correct. It did fail with mysterious
errors when I added some more commands into  the shell_commands.