Searialize channel readability setting
This is the major fix of this branch. The Netty framework actually
does not guarantee the same order of actually setting the readbility
of a channel as the order of the setReadability method call when
called sequentially by different Netty worker threads.  This problem
may cause unexpected delayed read pause, which is the main cause of
query hang.

Implement the sequential actual readability setting.