MasterDaemonTest: catch and throw the exceptiona instead of timeout.
If an exception was thrown in try{}, such as an AssertionError, it needs
to be exposed. Before this commit, we were simply waiting for the port
to be released in finally{} which would never happen since shutdown was
never called. To do that, catch the exception, shutdown the master
daemon, and put thread checks in finally{}.