Fix two things:
1. if a LocalFragment has the correct state EXECUTION_READY,
STATE_EXECUTION_REQUESTED will be added in execute(), then the task will
be submitted and a thread will be running executeActually(). However, if
the state of this fragment changed between these two timestamps, for
example, output was disabled, then when executeActually() is called the
fragment won't be actually executed. In this case we need to unset
STATE_EXECUTION_REQUESTED, since execute() checks if the state equals to
EXECUTION_READY. If STATE_EXECUTION_REQUESTED is still there then the
task won't be submitted by execute() anymore, even if the output is
enabled later.

2. if a physical channel disconnected, we also need to unmap its logical
channels. This should only happen when there is a failure.