SingleGroupByAggregate: keep the pipeline full if used as combiner
When aggregates are used as combiners, we can actually empty them and send the state along
whenever. The disadvantage is extra tuples sent over the network; the advantage is lower
memory consumption and potentially faster query completion because the pipeline is kept
full. Add a simple version of this pipeline-preserving behavior.