profiler: prep for adding subquery profiling
Add a subQueryId column to the database for profiling events and sending.
(It was already there for resources, so I did not touch ANY resource-related
code in the codebase.)

Then propagate query and subquery-aware changes across the codebase -- namely
in the indexes, queries, and REST API.

Will need to alter table for existing profiling data.