LeapFrogJoin: gracefully handle case when tables==null
Fix #689.

My belief is that the following happens: a fragment reaches EOS before
the entire query does. In this case, the root operator calls cleanup()
on its children, which leads to LFJ nulling out the .tables field.
However, the resource profiling code still measures the hash table
size, and thus was triggering an NPE.